refactor(recipes): _run_validation_epoch → BaseRecipe (PP-aware for VLM)
VLM adopts LLM's pattern: engine.forward_backward(num_label_tokens=None)
sums unnormalized batch losses, dp_allreduce, then divides by the total
label-token count. Under PP, the value travels from the last pipeline
stage to rank 0 so whichever rank logs has the data.

The shared method lives on BaseRecipe. Both recipes' duplicates removed.
VLM gains PP-aware validation; the prior "Validation is not supported
for pipeline parallelism" warn-and-skip behavior is dropped.
